+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- support ack_timeout None value
+- when killing a talker job wait until process is terminated (remove hanging job mechanism)
+- handle properly corrupted seen jobs file
 
 ## [1.9.0] - 2020-03-05
 ### Fixed